refactor: (#11) 방 생성 모달 index.tsx로 통합
