+++ src/features/main/ui/CreateRoomModal/MaxPlayerSelector.tsx
