--- src/features/main/ui/CreateRooms/index.ts
@@ -1,7 +1,0 @@
-export { CreateRoomModal } from './CreateRoomModal';
-export { CreateConfirmButton } from './CreateConfirmButton';
-export { GameModeSelector } from './GameModeSelector';
-export { MaxPlayerSelector } from './MaxPlayerSelector';
-export { PasswordInput } from './PasswordInput';
-export { RoomTitleInput } from './RoomTitleInput';
-export { VisibilitySelector } from './VisibilitySelector';
